pyproject.toml
[project]
name = "clickup-stories-mcp"
version = "0.1.0"
description = "MCP server for ClickUp task management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
clickup-mcp = "clickup_mcp.server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"